perf: tokenizer zero-alloc, AVX2 hadd reduction, bbox pre-filter e spin-poll calibrado

+++ go_api/engine/engine.go
@@ -1,7 +1,7 @@
 package engine
 
 //go:noescape
-func SearchVectorFast(q *float32, scratch *byte) int32
+func SearchVectorFast(q *int16, scratch *byte) int32
 
 //go:noescape
 func Pause()
